+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/dao/common/StoreErrorCodeInfoDao.kt
@@ -66,7 +66,7 @@ class StoreErrorCodeInfoDao {
         }).execute()
     }
 
-    fun getAtomErrorCode(
+    fun getStoreErrorCode(
         dslContext: DSLContext,
         storeCode: String,
         storeType: StoreTypeEnum,
