+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/dao/common/StoreStatisticDailyDao.kt
@@ -179,7 +179,7 @@ class StoreStatisticDailyDao {
         }
     }
 
-    fun getAtomExecuteCountByCode(
+    fun getStoreExecuteCountByCode(
         dslContext: DSLContext,
         storeCode: String,
         storeType: Byte,
