+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/resources/atom/ServiceAtomResourceImpl.kt
@@ -33,20 +33,16 @@ import com.tencent.devops.store.pojo.atom.AtomClassifyInfo
 import com.tencent.devops.store.pojo.atom.AtomProp
 import com.tencent.devops.store.pojo.atom.InstalledAtom
 import com.tencent.devops.store.pojo.atom.PipelineAtom
-import com.tencent.devops.store.pojo.common.enums.ErrorCodeTypeEnum
-import com.tencent.devops.store.pojo.common.enums.StoreTypeEnum
 import com.tencent.devops.store.service.atom.AtomPropService
 import com.tencent.devops.store.service.atom.AtomService
 import com.tencent.devops.store.service.atom.MarketAtomClassifyService
-import com.tencent.devops.store.service.atom.MarketAtomErrorCodeService
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
 class ServiceAtomResourceImpl @Autowired constructor(
     private val atomService: AtomService,
     private val atomPropService: AtomPropService,
-    private val atomClassifyService: MarketAtomClassifyService,
-    private val marketAtomErrorCodeService: MarketAtomErrorCodeService
+    private val atomClassifyService: MarketAtomClassifyService
 ) : ServiceAtomResource {
 
     override fun getInstalledAtoms(
@@ -73,21 +69,5 @@ class ServiceAtomResourceImpl @Autowired constructor(
 
     override fun getAtomClassifyInfo(atomCode: String): Result<AtomClassifyInfo?> {
         return atomClassifyService.getAtomClassifyInfo(atomCode)
-    }
-
-    override fun isComplianceErrorCode(
-        storeCode: String,
-        storeType: StoreTypeEnum,
-        errorCode: Int,
-        errorCodeType: ErrorCodeTypeEnum
-    ): Result<Boolean> {
-        return Result(
-            marketAtomErrorCodeService.isComplianceErrorCode(
-                storeCode,
-                storeType,
-                errorCode,
-                errorCodeType
-            )
-        )
     }
 }
