+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/resources/common/ServiceStoreResourceImpl.kt
@@ -33,7 +33,9 @@ import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.store.api.common.ServiceStoreResource
 import com.tencent.devops.store.pojo.common.SensitiveConfResp
 import com.tencent.devops.store.pojo.common.StoreBuildResultRequest
+import com.tencent.devops.store.pojo.common.enums.ErrorCodeTypeEnum
 import com.tencent.devops.store.pojo.common.enums.StoreTypeEnum
+import com.tencent.devops.store.service.atom.MarketAtomErrorCodeService
 import com.tencent.devops.store.service.common.StoreBuildService
 import com.tencent.devops.store.service.common.StoreMemberService
 import com.tencent.devops.store.service.common.StoreProjectService
@@ -44,7 +46,8 @@ import org.springframework.beans.factory.annotation.Autowired
 class ServiceStoreResourceImpl @Autowired constructor(
     private val storeProjectService: StoreProjectService,
     private val sensitiveConfService: UserSensitiveConfService,
-    private val storeBuildService: StoreBuildService
+    private val storeBuildService: StoreBuildService,
+    private val marketAtomErrorCodeService: MarketAtomErrorCodeService
 ) : ServiceStoreResource {
 
     override fun uninstall(storeCode: String, storeType: StoreTypeEnum, projectCode: String): Result<Boolean> {
@@ -69,6 +72,22 @@ class ServiceStoreResourceImpl @Autowired constructor(
                 clazz = StoreMemberService::class.java,
                 beanName = "${storeType.name.toLowerCase()}MemberService"
             ).isStoreMember(userId, storeCode, storeType.type.toByte())
+        )
+    }
+
+    override fun isComplianceErrorCode(
+        storeCode: String,
+        storeType: StoreTypeEnum,
+        errorCode: Int,
+        errorCodeType: ErrorCodeTypeEnum
+    ): Result<Boolean> {
+        return Result(
+            marketAtomErrorCodeService.isComplianceErrorCode(
+                storeCode,
+                storeType,
+                errorCode,
+                errorCodeType
+            )
         )
     }
 }
