+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/atom/impl/MarketAtomErrorCodeServiceImpl.kt
@@ -132,7 +132,7 @@ class MarketAtomErrorCodeServiceImpl @Autowired constructor(
                 ) != null
             } ?: false
         }
-        return storeErrorCodeInfoDao.getAtomErrorCode(
+        return storeErrorCodeInfoDao.getStoreErrorCode(
             dslContext = dslContext,
             storeCode = storeCode,
             storeType = storeType,
