+++ support-files/sql/1001_ci_metrics_ddl_mysql.sql
@@ -230,20 +230,20 @@ CREATE TABLE IF NOT EXISTS `T_ERROR_CODE_INFO` (
   `ATOM_CODE` varchar(64) DEFAULT NULL COMMENT '关联插件Code',
   PRIMARY KEY (`ID`),
   UNIQUE KEY `T_ERROR_CODE_INFO_UN` (`ERROR_TYPE`,`ERROR_CODE`,`ATOM_CODE`),
-  KEY `UNI_TECI_CODE` (`ERROR_CODE`),
-  KEY `UNI_TECI_ATOM_CODE` (`ATOM_CODE`)
+  KEY `INX_TECI_CODE` (`ERROR_CODE`),
+  KEY `INX_TECI_ATOM_CODE` (`ATOM_CODE`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='错误码信息表';
 
 CREATE TABLE IF NOT EXISTS `T_ATOM_INDEX_STATISTICS_DAILY` (
   `ID` bigint(32) NOT NULL COMMENT '主键ID',
   `ATOM_CODE` varchar(64) NOT NULL COMMENT '插件代码',
-  `FAIL_COMPLIANCE_COUNT` int(11) NOT NULL DEFAULT '0' COMMENT '插件错误合规次数',
+  `FAIL_COMPLIANCE_COUNT` int(11) NOT NULL DEFAULT 0 COMMENT '插件错误合规次数',
   `STATISTICS_TIME` datetime(3) NOT NULL COMMENT '统计时间',
   `CREATOR` varchar(50) NOT NULL DEFAULT 'system' COMMENT '创建者',
   `MODIFIER` varchar(50) NOT NULL DEFAULT 'system' COMMENT '修改者',
   `UPDATE_TIME` datetime(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3) COMMENT '修改时间',
   `CREATE_TIME` datetime(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3) COMMENT '创建时间',
-  `FAIL_EXECUTE_COUNT` int(11) NOT NULL DEFAULT '0' COMMENT '失败执行次数',
+  `FAIL_EXECUTE_COUNT` int(11) NOT NULL DEFAULT 0 COMMENT '失败执行次数',
   PRIMARY KEY (`ID`),
   UNIQUE KEY `UNI_TAISD_ATOM_TIME` (`ATOM_CODE`,`STATISTICS_TIME`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COMMENT='插件每日指标数据统计表';
